Added Exception Handling for Card Details

+++ src/main/java/com/b3t3/loanAdminManagement/service/Loan_Card_Master_service_impl.java
@@ -3,6 +3,9 @@ package com.b3t3.loanAdminManagement.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
+
+import com.b3t3.loanAdminManagement.Exception.IdAlreadyExistsException;
+import com.b3t3.loanAdminManagement.Exception.IdDoesNotExistException;
 import com.b3t3.loanAdminManagement.dao.Loan_Card_Master_dao;
 import com.b3t3.loanAdminManagement.model.Loan_Card_Master;
 
@@ -11,9 +14,9 @@ public class Loan_Card_Master_service_impl implements Loan_Card_Master_service{
 	@Autowired
 	Loan_Card_Master_dao dao;
 	
-	public String addCard(Loan_Card_Master card) {
+	public String addCard(Loan_Card_Master card) throws IdAlreadyExistsException {
 		if(dao.existsById(card.getLoan_id())) {
-			return "Card Already Present";
+			throw new IdAlreadyExistsException();
 		}else {
 			dao.save(card).toString();
 			return "Card Added Successfully!";
@@ -24,14 +27,22 @@ public class Loan_Card_Master_service_impl implements Loan_Card_Master_service{
 		return (List<Loan_Card_Master>)dao.findAll();
 	}
 	
-	public String deleteCard(String id) {
-		dao.deleteById(id);
-		return "Id " + id + " deleted";
+	public String deleteCard(String id) throws IdDoesNotExistException {
+		if(!dao.existsById(id))
+			throw new IdDoesNotExistException();
+		else {
+			dao.deleteById(id);
+			return "Id " + id + " deleted";
+		}
 	}
 
-	public String updateCard(Loan_Card_Master update_card) {
-		dao.save(update_card);
-		return "Card updated successfully";
+	public String updateCard(Loan_Card_Master update_card) throws IdDoesNotExistException {
+		if(!dao.existsById(update_card.getLoan_id()))
+			throw new IdDoesNotExistException();
+		else {
+			dao.save(update_card);
+			return "Card updated successfully";
+		}
 	}
